Refactor #239: Moved graceful shutdown logic and signal handling into a dedicated utility module
